+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/DateTimeBuilder.java
@@ -1,11 +1,16 @@
 package org.enso.table.data.column.builder;
 
+import java.time.LocalDate;
+import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import org.enso.table.data.column.storage.Storage;
+import org.enso.table.data.column.storage.datetime.DateStorage;
 import org.enso.table.data.column.storage.datetime.DateTimeStorage;
 import org.enso.table.data.column.storage.type.DateTimeType;
+import org.enso.table.data.column.storage.type.DateType;
 import org.enso.table.data.column.storage.type.StorageType;
 import org.enso.table.error.ValueTypeMismatchException;
+import org.graalvm.polyglot.Context;
 
 /** A builder for string columns. */
 public class DateTimeBuilder extends TypedBuilderImpl<ZonedDateTime> {
@@ -23,12 +28,48 @@ public class DateTimeBuilder extends TypedBuilderImpl<ZonedDateTime> {
     return DateTimeType.INSTANCE;
   }
 
+  /**
+   * TODO DRY {@link org.enso.table.data.column.operation.cast.ToDateTimeStorageConverter}
+   * convertDate.
+   */
+  private ZonedDateTime convertDate(LocalDate date) {
+    return date.atStartOfDay().atZone(ZoneId.systemDefault());
+  }
+
   @Override
   public void appendNoGrow(Object o) {
     try {
       data[currentSize++] = (ZonedDateTime) o;
     } catch (ClassCastException e) {
       throw new ValueTypeMismatchException(getType(), o);
+    }
+  }
+
+  @Override
+  public void appendBulkStorage(Storage<?> storage) {
+    if (storage.getType() instanceof DateType) {
+      if (storage instanceof DateStorage dateStorage) {
+        Context context = Context.getCurrent();
+        for (int i = 0; i < dateStorage.size(); ++i) {
+          LocalDate date = dateStorage.getItemBoxed(i);
+          if (date == null) {
+            data[currentSize++] = null;
+          } else {
+            data[currentSize++] = convertDate(date);
+          }
+
+          context.safepoint();
+        }
+      } else {
+        throw new IllegalStateException(
+            "Unexpected storage implementation for type "
+                + storage.getType()
+                + ": "
+                + storage
+                + ". This is a bug in the Table library.");
+      }
+    } else {
+      super.appendBulkStorage(storage);
     }
   }