+++ std-bits/table/src/main/java/org/enso/table/data/table/Column.java
@@ -171,16 +171,15 @@ public class Column {
 
     Object converted = Polyglot_Utils.convertPolyglotValue(item);
 
-    Builder builder;
     if (converted == null) {
-      builder = new MixedBuilder(repeat);
-    } else {
-      StorageType storageType = StorageType.forBoxedItem(converted);
-      builder = Builder.getForType(storageType, repeat, problemAggregator);
+      Builder builder = new MixedBuilder(repeat);
+      builder.appendNulls(repeat);
+      return new Column(name, builder.seal());
     }
 
+    StorageType storageType = StorageType.forBoxedItem(converted);
+    Builder builder = Builder.getForType(storageType, repeat, problemAggregator);
     Context context = Context.getCurrent();
-
     for (int i = 0; i < repeat; i++) {
       builder.appendNoGrow(converted);
       context.safepoint();